requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_pytorch_select=1.1.0=cpu
blas=1.0=mkl
ca-certificates=2019.8.28=0
certifi=2019.9.11=py37_0
cffi=1.12.3=py37h7a1dbc1_0
click=7.0=py37_0
cycler=0.10.0=py37_0
flask=1.1.1=py_0
freetype=2.9.1=ha9979f8_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
intel-openmp=2019.4=245
itsdangerous=1.1.0=py37_0
jinja2=2.10.1=py37_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.1.0=py37ha925a31_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.0.10=hb898794_2
markupsafe=1.1.1=py37he774522_0
matplotlib=3.1.1=py37hc8f65d3_0
mkl=2019.4=245
mkl-service=2.0.2=py37he774522_0
mkl_fft=1.0.14=py37h14836fe_0
mkl_random=1.0.2=py37h343c172_0
ninja=1.9.0=py37h74a9793_0
numpy=1.16.4=py37h19fb1c0_0
numpy-base=1.16.4=py37hc3f5095_0
olefile=0.46=py37_0
openssl=1.1.1c=he774522_1
pandas=0.25.0=py37ha925a31_0
patsy=0.5.1=py37_0
pillow=6.1.0=py37hdc69c19_0
pip=19.1.1=py37_0
pycparser=2.19=py37_0
pyparsing=2.4.2=py_0
pyqt=5.9.2=py37h6538335_2
python=3.7.4=h5263a28_0
python-dateutil=2.8.0=py37_0
pytorch=1.0.1=cpu_py37h39a92a0_0
pytz=2019.2=py_0
qt=5.9.7=vc14h73c81de_0
scipy=1.3.1=py37h29ff71c_0
seaborn=0.9.0=py37_0
setuptools=41.0.1=py37_0
sip=4.19.8=py37h6538335_0
six=1.12.0=py37_0
sqlite=3.29.0=he774522_0
statsmodels=0.10.1=py37h8c2d366_0
tk=8.6.8=hfa6e2cd_0
torchvision=0.2.2=py_3
tornado=6.0.3=py37he774522_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.15.26706=h3a45250_4
werkzeug=0.16.0=py_0
wheel=0.33.4=py37_0
wincertstore=0.2=py37_0
xz=5.2.4=h2fa13f4_4
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
